Remove unneeded main_hub_text variant needle
This somehow snuck in (in the wrong directory) with the ppc64
merges, but it's not needed. I've verified with the admin tab
for checking on needle use; this needle has never been matched,
and all the ppc64 text install tests match the non-variant
needle just fine. So I'm removing this unneeded variant.