Add variant device_sda_selected needle for #1530410
Rawhide has a bit of a problem where its 'description' of an
iSCSI disk is so long that the other columns that should appear
in the CONFIGURE MOUNT POINT dialog don't. This means our
device_sda_selected needle doesn't match, because the column
where 'sda' should appear isn't visible.

So add a soft-fail needle to cover this case; we know what the
description for the disk that's 'sda' in this case looks like,
so match on that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>