Narrow device_type needle area to fix match
In the most recent Rawhide compose, anaconda has changed how it
names volume groups (again), and now the VGs in openQA installs
have rather longer names. This winds up causing a problem for
this needle: the column where the 'Device Type' dropdown is
placed actually gets narrower (because the column to the right,
which has the volume group name in it, is wider), and so the
dropdown box is narrower, and the arrow protrudes into the area
which the needle expects not to have an arrow in it. So, let's
make the match area slightly narrower.

Signed-off-by: Adam Williamson <awilliam@redhat.com>