Remove install-no-user for F27 live till I can make it work
This doesn't really work, it needs a product with that explicit
version to make it work, apparently. I think I'd better just
take this out and read up again on how the wildcards work rather
than just messing around with it. I'll put it back if I can be
reasonably sure of making it work.

Signed-off-by: Adam Williamson <awilliam@redhat.com>