Add Workstation upgrade test to update tests
Since we set up this whole process to run upgrade tests for
updates to run the FreeIPA ones on the Server base image, it
should be easy to also run Workstation upgrade tests on the
Workstation base image. So let's do that! Let's do the most
complex test only, the encrypted upgrade one.

Signed-off-by: Adam Williamson <awilliam@redhat.com>