Switch to std graphics to avoid vt/splash issues
Time for the great video driver merry-go-round again...lately
we're having some issues with vt corruption and rendering over
the boot splash, let's try using 'std' driver to see if it
helps a bit. stg has been set up this way for a couple of days
and it's not blowing up at least.

Signed-off-by: Adam Williamson <awilliam@redhat.com>