Handle rename of AtomicWorkstation to Silverblue
We updated the dispatcher code, but not the templates...so we
didn't have any actual tests to run on Silverblue images. Let's
fix that. Note this means we won't test F28 and earlier images
that still have 'AtomicWorkstation' as their subvariant any
more, unless I set up some sorta workaround in the dispatcher.

Signed-off-by: Adam Williamson <awilliam@redhat.com>