Add new-style UEFI firmware vars to machine templates
We're about to deploy recent os-autoinst to staging. The UEFI
stuff got rewritten since the version we currently have, and
this is now the recommended way to handle UEFI firmware. The
new vars shouldn't confuse the old code still deployed on prod,
it should just ignore these vars and keep working off the old
'UEFI=1' (which causes it to autodetect the file locations).

Signed-off-by: Adam Williamson <awilliam@redhat.com>